backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7

# Environment and Configuration
python-decouple==3.8

# Cloud Storage
django-storages==1.14.2
boto3==1.29.7
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Image Processing
Pillow==10.0.1

# Video Processing
opencv-python==4.8.1.78

# File Handling
python-magic==0.4.27

# Email
django-ses==3.5.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Security
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5